Fixed #81
Added code for showing/editing the locations/rooms.  This works the same
way as the code for showing/editing time slots.  There are a couple of
differences worth noting, and a separate commit will bring the slot
editing to match this.  The differences are:
	1) There was inconsistent use of "input type=button" v "button".
Since none of the buttons represent form data or submit or reset the
form, "button" should be preferred.
	2) It seems that some browsers add the legend tag into the tab
order.  This sets "tabindex=-1" for the legend, so that you don't have
to tab to or past it when navigating with the keyboard.  I found this
one by accident. :)