Fixed #66
This really needs far more testing, as I've only tested a simple case by resetting the leader for workshops "topic13" and "topic4" to be "user53", and ensuring we have 3 rounds and 4 rooms.

The new code attempts to look for conflicts after assigning everything, (note that a conflict doesn't count if there is more than one leader for a workshop).  It then attempts to switch a conflicting workshop with another (but never switching with something assigned to location_id 0 or the prep BOF).  It then tests for conflicts again.  If there are fewer conflicts, this change is commited to the DB, and the next conflict is addressed.  If there are the same or more conflicts, the change is rolled back, and it attempts to move a different workshop for the same conflict. This continues until all conflicts are resolved, or a maximum number of attempts has been reached.
One of the areas of testing that I haven't focused on, is what happens if the conflict is with something in location 0 (or the prep BOF) to begin with.  In that case, it might well attempt to switch something from location 0, which would be less than ideal.  Tweaking the queries might prevent such an occurrence.