Split functionality of Results.php into two more new classes, DBO and Logger.
All SQL queries were moved into the new DBO class.  This class exists
only to run SQL queries on the database.  This class does not maintain
state data, EXCEPT for a member which holds a reference to the database
PDO object and the ID of the Prep BoF.  Some functions use static
members to hold their query objects -- this is done to improve
performance in unit tests, as it's unlikely to have much impact on a
real running system.  Some performance decrease is expected because of
this, but it should be negligible, and possibly not even measurable.
For some methods, there might be a very slight performance increase, but
again, it should be negligible.  Performance testing in a real scenario
was not performed, so I might be wrong. :)

The new Logger class is used to handle the logging facilities of
Results.php.  Some methods expect to receive a DBO object so they can
look up names for prettier logging.  Some of the methods in the DBO
class expect to receive a Logger object, too.  Neither class holds a
reference to the other.  The Logger class may cache location and round
names from the DBO object.  This would be a problem if the Logger object
is kept alive through admin changes, like the number and names of
locations and rounds.  It's not anticipated this will be a real problem.
If it is, not caching those names would be simplest, and probably
perform well enough.  One could argue that such caching is a premature
optimization, since no performance testing was done; it's cached only
because the original code in Results.php cached those names.

Results.php is passed a new Logger object every time the Results object
is created.  The DBO object is handled as a singleton in the app's
container (see changes to dependencies.php), and is passed in every time
the Results object is created.

This new architecture makes it easier to create and run unit tests.
Unit tests for the Results class can mock out the database and any
logging entirely, while still exercising all of its logic.  The Logger
class can likewise be tested, mocking out any actual database
dependencies (like location and round names).  Finally, the DBO class's
unit test can contain complex logic for setting up the database in
predictable patterns and ensure that data is returned or modified as
expected.

Important note: Because the resolveConflicts logic relies on being able
to rollback transactions, the ENGINE for the workshop table MUST be
InnoDB.  The sql/createtables.sql file has been updated to reflect this.
To use an existing database, run the command:

	ALTER TABLE workshop ENGINE=InnoDB;

on the database.  Otherwise, if there are conflicts, really weird things
will probably happen. :)

Re-ordered code so private methods appear before public methods, and
methods are alphabetized.

Added doc comments.

The next commit will contain the actual unit tests.