Added unit tests for Auth
This involved creating a new Cookies class for setting cookies (we can't
test any method which calls setcookies), moving code into DBO.php, and
passing in the Translator object, instead of using it as a global.

Added tests for new functions in DBO.php.

Updated how Auth object is created, passing in DBO, an instance of the
new Cookies class, and the translator object.

Added require for Cookies.php to public/index.php