Moved SQL from Admin.php into DBO.php and added unit tests for it
Also adjusted how Admin.php creates/uses the Results class (it's not
created in dependencies.php and passed into Admin, not created
directly).  This change also meant it was reasonable to remove display code
from the Results class.  The Results class now just returns the results,
and lets the Admin class display it.

Because the admin class invokes mysqldump for its "download database"
functionality, php-mock was added to the dependencies to prevent actually
running mysqldump.