Added XML coverage output
This might be useful for Codacy.  It's added to this pull request to see
what Codacy does with it; coverage was uploaded to Codacy for the
franc6/BOF repository prior to this commit.  By updating the pull
request, Codacy might generate a new report and show coverage for the
pull request.  At least, that's my hope. :)