Added more cypress tests
Some are end-to-end, others only test the UI.

Updated src/classes/Admin.php to ensure Content-Type header is set when
downloading the database; this also allows for an integration test for
it.

Removed cypress/support/* from .gitignore, since the files are something
we need to preserve, now that there are changes to them.  We might need
to be sure the installation process doesn't overwrite them.