Modified code to allow not scheduling the Prep BoF
Modified DBO::getPrepBoF so it can return false if a new configuration
option is set.

Modified Results::bookPrepBoF to return true if it scheduled a Prep BoF,
or false if not (based on the return value of DBO::getPrepBoF).  Also
modified it to schedule the Prep BoF for the LAST room.

Updated Results::fillBooking to schedule an additional workshop if the
PrepBof was not scheduled.

Updated Results::bookWorkshops to check the return value of bookPrepBoF,
and pass it along to fillBooking, so its changes work.

Updated all tests for these changes.