Added backend support for explicitly setting the location and round for the Prep BoF
Also changed the UI some to handle differences in the back-end from the
original expectation.

Added phpunit tests for the new functionality.

UI tests still must be added.

There are two funny things going on here:

1) "Default" is added to the end of the UI as an explicit default
setting for both location and round.  This prevents weirdness with
adding/deleting rounds/locations -- if "Default" was set, then adding
and deleting rounds/locations won't change that.

2) Because PHP's intval() returns 0 on errors, we need to reserve 0 as
an error condition.  This means that when storing data into the
database, we have to use a 1-based index.  All API calls assume a
0-based index which should be more intuitive for developers.  The
1-based index only affects storage; the DBO class automatically adjusts
between 0-based and 1-based indices.