Added integration tests for new prep bof functionality
These include pure-UI tests as well as end-to-end tests.

The new end-to-end tests provide some example and framework for adding
more end-to-end tests to check DB values after making changes.

Updated Cypress to version 5 (version 7 is the latest, but it looks like
6 & 7 will require some invasive changes to existing tests, so updates
to 6 or 7 should be their own commit and PR).  This also required adding
typescript to the dependencies.

Running npm update (when updating cypress) also updated a few other
packages by a minor revision.